more rails changes

+++ gen/rails/RailsGen.java
@@ -242,7 +242,7 @@ public class RailsGen extends Generator {
             HTMLUtils.addDiv(buf, "form-group");
             StringUtils.addLine(buf, "<input type=\"text\" class=\"form-control\" id=\"search\" name=\"search\" placeholder=\"Search\">");
             HTMLUtils.closeDiv(buf);
-            StringUtils.addLine(buf, "<button type=\"submit\" class=\"btn btn-success\">Search</button>");
+            StringUtils.addLine(buf, "<button type=\"submit\" class=\"btn btn-primary\">Search</button>");
             StringUtils.addLine(buf, "</form>");
         }
 
@@ -541,13 +541,14 @@ public class RailsGen extends Generator {
                     line += "<td><%= number_to_currency(" + modelName + "." + fieldName + ", :unit => \"$\") %></td>";
                 else if (fType.equals(Type.IMAGE)) {
                     StringBuilder imgBuilder = new StringBuilder("<td>");
-                    HTMLUtils.addRuby(imgBuilder, "if " + modelName + ".url.nil?"); // TODO remove hardcoding
+                    HTMLUtils.addRuby(imgBuilder, "if " + modelName + "." + fieldName + ".url.nil?"); // TODO remove hardcoding
                     HTMLUtils.addRubyOutput(imgBuilder, "image_tag(\"ImagePlaceholderSmall.png\")");
                     HTMLUtils.addRuby(imgBuilder, "else");
                     //<%= image_tag micropost.picture.url if micropost.picture? %>
                     HTMLUtils.addRubyOutput(imgBuilder, "image_tag(" + modelName + "." + fieldName + ".url, :size => \"50x50\", :style => \"width: 50px; height: 50px\")");
                     HTMLUtils.addRuby(imgBuilder, "end");
                     imgBuilder.append("</td>");
+                    line += imgBuilder.toString();
                 }
                 else if (fType.equals(Type.URL))
                     line += "<td><%= " + modelName + "." + fieldName + " %></td>";
@@ -658,6 +659,51 @@ public class RailsGen extends Generator {
         HTMLUtils.addDiv(buf, "form-group");
         HTMLUtils.addRubyOutput(buf, "f.label(:" + name + ", class: \"col-sm-2 control-label\") ");
         HTMLUtils.addDiv(buf, "col-sm-8");
+        HTMLUtils.addRubyOutput(buf, "f." + fieldType + "_field(:" + name + ", class: \"form-control\") ");
+        HTMLUtils.closeDiv(buf);
+        HTMLUtils.closeDiv(buf);
+    }
+
+    private void generateStaticFormField (StringBuilder buf, String modelName, String fieldName) {
+        HTMLUtils.addDiv(buf, "form-group");
+        HTMLUtils.addRubyOutput(buf, "f.label(:" + fieldName + ", class: \"col-sm-2 control-label\") ");
+        HTMLUtils.addDiv(buf, "col-sm-8");
+        HTMLUtils.addParagraph(buf, "<%= @" + modelName + "." + fieldName + " %>", "form-control-static");
+        HTMLUtils.closeDiv(buf);
+        HTMLUtils.closeDiv(buf);
+    }
+
+    private void generateCurrencyField (StringBuilder buf, String name) {
+        HTMLUtils.addDiv(buf, "form-group");
+        HTMLUtils.addRubyOutput(buf, "f.label(:" + name + ", class: \"col-sm-2 control-label\") ");
+        HTMLUtils.addDiv(buf, "col-sm-8");
+        HTMLUtils.addDiv(buf, "input-group");
+        HTMLUtils.addDivWithContent(buf, "input-group-addon", "$");
+        HTMLUtils.addRubyOutput(buf, "f.text_field(:" + name + ", class: \"form-control\") ");
+        HTMLUtils.closeDiv(buf);
+        HTMLUtils.closeDiv(buf);
+        HTMLUtils.closeDiv(buf);
+    }
+
+    private void generateDurationField (StringBuilder buf, String name, Type subType) {
+        HTMLUtils.addDiv(buf, "form-group");
+        HTMLUtils.addRubyOutput(buf, "f.label(:" + name + ", class: \"col-sm-2 control-label\") ");
+        HTMLUtils.addDiv(buf, "col-sm-8");
+        String typeName = subType == null ? "Minutes:Hours" : subType.getName();
+        HTMLUtils.addRubyOutput(buf, "f.text_field(:" + name + ", class: \"form-control\", placeholder:  \"" + typeName + "\") ");
+        HTMLUtils.closeDiv(buf);
+        HTMLUtils.closeDiv(buf);
+    }
+
+    private void generateRangeFields (StringBuilder buf, String name, Type subType) {
+        HTMLUtils.addDiv(buf, "form-group");
+        HTMLUtils.addRubyOutput(buf, "f.label(:" + name + ", class: \"col-sm-2 control-label\") ");
+        String fieldType = "text"; // TODO switch on subType
+        HTMLUtils.addDiv(buf, "col-sm-3");
+        HTMLUtils.addRubyOutput(buf, "f." + fieldType + "_field(:" + name + ", class: \"form-control\") ");
+        StringUtils.addLine(buf, "...");
+        HTMLUtils.closeDiv(buf);
+        HTMLUtils.addDiv(buf, "col-sm-3");
         HTMLUtils.addRubyOutput(buf, "f." + fieldType + "_field(:" + name + ", class: \"form-control\") ");
         HTMLUtils.closeDiv(buf);
         HTMLUtils.closeDiv(buf);
@@ -911,6 +957,7 @@ public class RailsGen extends Generator {
 //               " *= require jquery-ui", true, true);
 
         FileUtils.copyTextFile("C:/Users/jsanchez/Downloads/apps/resources/simple2.css.scss", app.getWebAppDir() + "/app/assets/stylesheets/custom.css.scss");
+        FileUtils.copyFile("C:/Users/jsanchez/Downloads/apps/resources/ImagePlaceholderSmall.png", app.getWebAppDir() + "/app/assets/images/ImagePlaceholderSmall.png");
         if (app.getJumbotronImageUrl() != null)
             addStyle(new String[] {".jumbotron {", "\tbackground-image: url('" + app.getJumbotronImageUrl() + "');", "}"});
 
@@ -921,7 +968,7 @@ public class RailsGen extends Generator {
 
             // TODO: full navbar, buttons, tables, forms, header, footer
             overrideStyles(new String [] {"jumbotron-color", "jumbotron-heading-color", "body-bg", "text-color", "link-color", "brand-primary", "brand-info", "navbar-inverse"},
-                new String [] {scheme.getColor1(), scheme.getColor1(), scheme.getColor2(), scheme.getColor3(), scheme.getColor3(), scheme.getColor4(), scheme.getColor4(), scheme.getColor5()});
+                new String [] {scheme.getFirstAccent(), scheme.getFirstAccent(), scheme.getBgPrimary(), scheme.getLettering(), scheme.getSecondAccent(), scheme.getFirstAccent(), scheme.getSecondAccent(), scheme.getBgSecondary()});
 
         }
         // TODO: not sure we need to do this if running with --without production:
@@ -1013,13 +1060,13 @@ public class RailsGen extends Generator {
                             tabbed(buf, "has_one :" + relModel.getName());
                         }
                         else if (relType.equals(RelType.MANY_TO_MANY)) {
-                            String  toAdd =  "has_many :" + relModel.getPluralName();
-                            if (rel.getThrough() != null) {
-                                toAdd += ", through: " + rel.getThrough().getPluralName();
-                            }
+                            String  toAdd;
+                            if (rel.getThrough() != null)
+                                toAdd =  "has_many :" + relModel.getPluralName() + ", through: " + rel.getThrough().getPluralName();
+                            else
+                                toAdd = "has_and_belongs_to_many :" + relModel.getPluralName();
                             tabbed(buf, toAdd);
                         }
-
                     }
                 }
                 if (model.hasImages()) {
@@ -1182,12 +1229,6 @@ public class RailsGen extends Generator {
         HTMLUtils.addDiv(buf, "page-header");
         HTMLUtils.addH1(buf, model.getCapName() + ": <%= @" + name + "." + model.getUserIndentifierFieldName() + " %>");
         HTMLUtils.closeDiv(buf);
-        /**
-         * <div class="page-header">
-         <h1>Example page header <small>Subtext for header</small></h1>
-         </div>
-         */
-
 
         StringBuilder       bodyContent = new StringBuilder();
 
@@ -1216,7 +1257,7 @@ public class RailsGen extends Generator {
                         generateReadOnlySection(bodyContent, "@" + nameFName, fName);
                     }
                     else {
-                        generateSublistView(bodyContent, name, fName, model.isSecure());
+                        generateSublistViewReadonly(bodyContent, name, fName);
                     }
                 }
                 else if (fName.equals("name")) {
@@ -1239,10 +1280,10 @@ public class RailsGen extends Generator {
                     generateReadOnlySection(bodyContent, "render_address(@" + nameFName + ") if @" + nameFName + "?", fName);
                 }
                 else if (fTypeName.equals(Type.URL.getName())) {
-                    StringUtils.addLine(bodyContent, "<% if @" + nameFName + "!= nil %>");
+                    //StringUtils.addLine(bodyContent, "<% if @" + nameFName + "!= nil %>");
                     String content  = "        <a href=\"<%= render_website(@" + nameFName + ") %>\" target=\"_blank\"><%= @" + nameFName +" %></a>";
-                    generateReadOnlySection(bodyContent, content, fName);
-                    StringUtils.addLine(bodyContent, "<% end %>");
+                    generateReadOnlySectionNoRuby(bodyContent, content, fName);
+                    //StringUtils.addLine(bodyContent, "<% end %>");
                 }
                 else if (fTypeName.equals(Type.BOOLEAN.getName())) {
                     generateReadOnlySection(bodyContent, "yesno(@" + nameFName + ")", fName);
@@ -1276,11 +1317,11 @@ public class RailsGen extends Generator {
                         break;
 
                     case ONE_TO_MANY:
-                        generateSublistView(bodyContent, name, fName, model.isSecure());
+                        generateSublistViewReadonly(bodyContent, name, fName);
                         break;
 
                     case MANY_TO_MANY:
-                        generateSublistView(bodyContent, name, fName, model.isSecure());
+                        generateSublistViewReadonly(bodyContent, name, fName);
                         break;
 
                     case MANY_TO_ONE:
@@ -1341,6 +1382,12 @@ public class RailsGen extends Generator {
         StringUtils.addLine(bodyContent, "</dd>");
     }
 
+    private void generateReadOnlySectionNoRuby (StringBuilder bodyContent, String fieldDetails, String fieldName) {
+        StringUtils.addLine(bodyContent, "<dt>" + WordUtils.capitalizeAndSpace(fieldName) + "</dt><dd>");
+        StringUtils.addLine(bodyContent, fieldDetails);
+        StringUtils.addLine(bodyContent, "</dd>");
+    }
+
     private void generateReadOnlyLinkSectionInForm (StringBuilder bodyContent, String fieldDisplayName, String fieldName, String linkPath) {
         HTMLUtils.addDiv(bodyContent, "form-group");
         StringUtils.addLine(bodyContent, "<label class=\"control-label col-sm-2\">" + fieldDisplayName + "</label>");
@@ -1351,7 +1398,7 @@ public class RailsGen extends Generator {
         StringUtils.addLine(bodyContent, "None");
         HTMLUtils.addRuby(bodyContent, "end");
         // TODO: this button needs to allow use to select another item - either list selection, lookup field, popup from a table select?
-        HTMLUtils.addButton(bodyContent, "change", "Change", "button");
+        HTMLUtils.addSmallButton(bodyContent, "change", "Change", "button");
         HTMLUtils.closeDiv(bodyContent);
         HTMLUtils.closeDiv(bodyContent);
     }
@@ -1368,8 +1415,23 @@ public class RailsGen extends Generator {
         //StringUtils.addLine(bodyContent, "</dd>");
     }
 
+    private void generateSublistViewReadonly (StringBuilder builder, String modelName, String collectionName) throws Exception {
+        String pluralColName = WordUtils.pluralize(collectionName);
+        StringUtils.addLine(builder, "<dl class=\"dl-horizontal\"><dt>" + WordUtils.capitalize(pluralColName) + "</dt><dd>");
+        StringUtils.addLine(builder, "<% if @" + modelName + "." + pluralColName + ".any? %>");
+
+        tabbed(builder, "<%= render @" + pluralColName + " %>");
+        tabbed(builder, "<% else %>");
+        tabbed(builder, "No " + pluralColName + " for this " + modelName + " yet.");
+        tabbed(builder, "<% end %>");
+
+        StringUtils.addLine(builder, "</dd></dl>");
+    }
+
     private void generateSublistView (StringBuilder builder, String modelName, String collectionName, boolean addAddBtn) throws Exception {
         String pluralColName = WordUtils.pluralize(collectionName);
+        StringUtils.addLine(builder, "<label class=\"control-label col-sm-2\">" + WordUtils.capitalize(pluralColName) + "</label>");
+        HTMLUtils.addDiv(builder, "col-sm-8");
         StringUtils.addLine(builder, "<% if @" + modelName + "." + pluralColName + ".any? %>");
 
         tabbed(builder, "<%= render @" + pluralColName + " %>");
@@ -1378,9 +1440,10 @@ public class RailsGen extends Generator {
         tabbed(builder, "<% end %>");
         if (addAddBtn) {
             tabbed(builder, "<% if logged_in? && current_" + modelName + "?(@" + modelName + ") %>");
-            tabbed(builder, "<%= link_to \"Add " + collectionName + "\", new_" + collectionName + "_path, class: \"btn btn-lg btn-primary\" %>");
+            tabbed(builder, "<%= link_to \"Add " + collectionName + "\", new_" + collectionName + "_path, class: \"btn btn-sm btn-default\" %>");
             tabbed(builder, "<% end %>");
         }
+        HTMLUtils.closeDiv(builder);
     }
 
     public void generateListView (Model model) throws Exception {
@@ -1477,6 +1540,11 @@ public class RailsGen extends Generator {
                      if (!(f.getName().equals("email") || f.getName().equals("name") || f.getName().equals("username") || f.getName().indexOf("password") != -1))
                          continue;
                  }
+                 else {
+                     // Don't typically allow editing these properties
+                     if (f.getName().indexOf("password") != -1)
+                         continue;
+                 }
 
                  String fName = f.getName();
 
@@ -1486,29 +1554,42 @@ public class RailsGen extends Generator {
                  if (fType.equals(Type.BOOLEAN)) {
                      generateCheckboxField(bodyContent, fName, capitalized + "?");
                  }
+                 else if (f.getName().equals("username")) {
+                     generateStaticFormField(bodyContent, model.getName(), fName);
+                 }
                  else if (fType.equals(Type.SHORT_STRING)) {
                      generateFormField(bodyContent, fName);
                  }
-                 else if (fType.equals(Type.LONG_STRING)) {
+                else if (fType.equals(Type.LONG_STRING)) {
                      generateTextArea(bodyContent, fName, 5);
                  }
                  else if (fType.equals(Type.STRING)) {
                      generateTextArea(bodyContent, fName, 2);
                  }
                  else if (fType.equals(Type.IMAGE)) {
-                     rubyout(bodyContent, "f.label :" + fName + ", \"Your " + WordUtils.capitalize(fName) + " Image\"");
+                     HTMLUtils.addDiv(bodyContent, "form-group");
+                     HTMLUtils.addRubyOutput(bodyContent, "f.label(:" + fName + ", class: \"col-sm-2 control-label\") ");
+                     HTMLUtils.addDiv(bodyContent, "col-sm-8");
+                     //rubyout(bodyContent, "f.label :" + fName + ", \"" + WordUtils.capitalize(fName) + " Image\"");
                      rubyout(bodyContent, "image_for(@" + fName + ")");
                      HTMLUtils.addDiv(bodyContent, "fileUpload btn");
-                     aline(bodyContent, "<span>Change Image</span>");
+                     //aline(bodyContent, "<span>Change Image</span>");
                      rubyout(bodyContent, "f.file_field :" + fName + ", :class => \"upload\"");
+                     HTMLUtils.closeDiv(bodyContent);
+                     HTMLUtils.closeDiv(bodyContent);
                      HTMLUtils.closeDiv(bodyContent);
                  }
                  else if (fType.equals(Type.PHONE)) {
                      generateFormField(bodyContent, fName, "phone");
-
+                 }
+                 else if (fType.equals(Type.DURATION)) {
+                     generateDurationField(bodyContent, fName, fType.getSubtype());
+                 }
+                 else if (fType.equals(Type.RANGE)) {
+                     generateRangeFields(bodyContent, fName, fType.getSubtype());
                  }
                  else if (fType.equals(Type.CURRENCY)) {
-                     generateFormField(bodyContent, fName);   // TODO: currency symbol?
+                     generateCurrencyField(bodyContent, fName);   // TODO: currency symbol?
                  }
                  else if (fType instanceof FixedList) {
 
@@ -1519,7 +1600,7 @@ public class RailsGen extends Generator {
                      HTMLUtils.addRubyOutput(bodyContent, "f.label(:" + fName + ", class: \"col-sm-2 control-label\") ");
                      HTMLUtils.addDiv(bodyContent, "col-sm-8");
 
-                     StringUtils.addLine(bodyContent, "<select name=\"" + name + "[" + fName + "]\" id=\"" + name + "_" + fName + "\" class=\"input-medium\">");
+                     StringUtils.addLine(bodyContent, "<select name=\"" + name + "[" + fName + "]\" id=\"" + name + "_" + fName + "\" class=\"form-control\">");
                      StringUtils.addLine(bodyContent, "<%= options_for_select ([");
 
                      if (values != null) {
@@ -2060,6 +2141,7 @@ public class RailsGen extends Generator {
 
                 if (rels != null) {
                     for (Rel rel : rels) {
+                        // TODO: do we need to generate belongs_to here instead in some cases?
                         tabbed(buf, "t.integer :" + rel.getModel().getName() + "_id", 3);
                     }
                 }
@@ -2073,7 +2155,10 @@ public class RailsGen extends Generator {
                 File f = FileUtils.fileEndingIn(app.getWebAppDir() + "/db/migrate/", endFileName);
 
                 if (f != null) {
-                    f.delete();
+                    if (FileUtils.textContentSame(buf.toString(), f))
+                        continue;
+                    else
+                        f.delete();
                 }
 
                 String fileName = String.valueOf(System.nanoTime()) + endFileName;
@@ -2250,8 +2335,11 @@ public class RailsGen extends Generator {
 
         builder = new StringBuilder();
         StringUtils.addLine(builder, "#!/bin/sh");
+        StringUtils.addLine(builder, "#This will show ruby processes to kill to stop server:");
+        StringUtils.addLine(builder, "tasklist |grep ruby | awk '{print $2}'");
         StringUtils.addLine(builder,  bundle + " exec rake db:drop");
         StringUtils.addLine(builder,  bundle + " exec rake db:migrate");
+        StringUtils.addLine(builder, "#Start server: rails.bat s");
 
         FileUtils.write(builder, app.getWebAppDir() + "/db.sh", true);
